requirements-dev.txt
black
fuzzywuzzy
isort
levenshtein
pytest-homeassistant-custom-component~=0.13.49
pytest
pytest-asyncio
pytest-cov
pycryptodome~=3.18
tinytuya~=1.12.11
aiohttp_cors